Figured out how DOM's work in P5.js
The text bar is not on top of the canvas thanks to literally thinking "outside the box" :P
Figured out the button problem form my proff (Thanks Adam!) But decided to use the other way cause it felt better interactively. (Didn't get much feedback from the other way."
Now to make a string to link to a png page and make some interactive comics! Oh and make the actual tool kit prettier.... and a little more functional